Change review columns from decimal to integer. Change review page to show wow factor. Change background color

+++ db/migrate/20140504214331_create_reviews.rb
@@ -1,9 +1,9 @@
 class CreateReviews < ActiveRecord::Migration
   def change
     create_table :reviews do |t|
-      t.decimal :patty
-      t.decimal :bun
-      t.decimal :bun_patty_ratio
+      t.integer :patty
+      t.integer :bun
+      t.integer :bun_patty_ratio
 
       t.timestamps
     end
